requirements.txt
Django==3.0.7
djangorestframework==3.11.1
django-cors-headers==3.1.1